+++ src/main/preload.ts
@@ -28,6 +28,7 @@ contextBridge.exposeInMainWorld('electronAPI', {
   getActiveTimeTrackingRecords: () => ipcRenderer.invoke('timeTracking:getActiveRecords'),
   getYesterdayTimeTrackingRecords: () => ipcRenderer.invoke('timeTracking:getYesterdayRecords'),
   getCurrentWeekTimeTrackingRecords: () => ipcRenderer.invoke('timeTracking:getCurrentWeekRecords'),
+  getWeekTimeTrackingRecords: (weekOffset: number) => ipcRenderer.invoke('timeTracking:getWeekRecords', weekOffset),
   updateTimeTrackingRecord: (record: TimeTrackingRecord) => ipcRenderer.invoke('timeTracking:updateRecord', record),
   deleteTimeTrackingRecord: (id: number) => ipcRenderer.invoke('timeTracking:deleteRecord', id),
   markAsUploaded: (id: number) => ipcRenderer.invoke('timeTracking:markAsUploaded', id),
@@ -90,6 +91,7 @@ declare global {
       getActiveTimeTrackingRecords: () => Promise<TimeTrackingRecord[]>;
       getYesterdayTimeTrackingRecords: () => Promise<TimeTrackingRecord[]>;
       getCurrentWeekTimeTrackingRecords: () => Promise<TimeTrackingRecord[]>;
+      getWeekTimeTrackingRecords: (weekOffset: number) => Promise<TimeTrackingRecord[]>;
       updateTimeTrackingRecord: (record: TimeTrackingRecord) => Promise<{ success: boolean }>;
       deleteTimeTrackingRecord: (id: number) => Promise<{ success: boolean }>;
       markAsUploaded: (id: number) => Promise<{ success: boolean }>;